requirements.txt
requests

streamlit>=1.33.0
langchain>=0.2.0
langchain-groq>=0.1.3
langchain-community>=0.2.0
python-dotenv>=1.0.1
yelpapi>=2.5.0
urllib3>=2.2.1
requests>=2.31.0
tqdm>=4.66.4
aiohttp>=3.9.5
tenacity>=8.2.3
